Moving useBindingSchema.tsx and useCurrentDataModelSchemaQuery.ts as well
